+++ examples/README.md
@@ -1,0 +1,7 @@
+EXAMPLES
+========
+
+These files are examples of how to use swarmkit. They are roughly maintained 
+and have some utility in testing changes to swarmkit, but should not be used 
+for other purposes. The official, recommended way to use swarmkit is as part of 
+Docker's Swarm Mode, which is built into Docker versions 1.12+.
