+++ examples/swarmctl/cluster/cmd.go
