+++ examples/swarmctl/cluster/common.go
