+++ examples/swarmctl/common/common.go
