+++ examples/swarmctl/common/print.go
