+++ examples/swarmctl/common/resolver.go
