+++ examples/swarmctl/network/common.go
