+++ examples/swarmctl/node/accept.go
