+++ examples/swarmctl/node/activate.go
