+++ examples/swarmctl/node/cmd.go
