+++ examples/swarmctl/node/demote.go
