+++ examples/swarmctl/node/drain.go
