+++ examples/swarmctl/node/pause.go
