+++ examples/swarmctl/node/promote.go
