+++ examples/swarmctl/node/remove.go
@@ -1,34 +1,37 @@
-package task
+package node
 
 import (
 	"errors"
-	"fmt"
 
 	"github.com/docker/swarmkit/api"
-	"github.com/docker/swarmkit/cmd/swarmctl/common"
+	"github.com/docker/swarmkit/examples/swarmctl/common"
 	"github.com/spf13/cobra"
 )
 
 var (
 	removeCmd = &cobra.Command{
-		Use:     "remove <task ID>",
-		Short:   "Remove a task",
+		Use:     "remove <node ID>",
+		Short:   "Remove a node",
 		Aliases: []string{"rm"},
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if len(args) == 0 {
-				return errors.New("task ID missing")
+				return errors.New("missing node ID")
 			}
+
 			c, err := common.Dial(cmd)
 			if err != nil {
 				return err
 			}
-
-			_, err = c.RemoveTask(common.Context(cmd), &api.RemoveTaskRequest{TaskID: args[0]})
+			node, err := getNode(common.Context(cmd), c, args[0])
 			if err != nil {
 				return err
 			}
-			fmt.Println(args[0])
-			return nil
+
+			_, err = c.RemoveNode(common.Context(cmd), &api.RemoveNodeRequest{
+				NodeID: node.ID,
+			})
+
+			return err
 		},
 	}
 )
