+++ examples/swarmctl/service/cmd.go
