+++ examples/swarmctl/service/common.go
