+++ examples/swarmctl/service/flagparser/mode.go
