+++ examples/swarmctl/service/flagparser/network.go
@@ -2,8 +2,8 @@ package flagparser
 
 import (
 	"github.com/docker/swarmkit/api"
-	"github.com/docker/swarmkit/cmd/swarmctl/common"
-	"github.com/docker/swarmkit/cmd/swarmctl/network"
+	"github.com/docker/swarmkit/examples/swarmctl/common"
+	"github.com/docker/swarmkit/examples/swarmctl/network"
 	"github.com/spf13/cobra"
 )
