+++ examples/swarmctl/service/flagparser/resource.go
