+++ examples/swarmctl/service/flagparser/restart.go
