+++ examples/swarmctl/service/remove.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/docker/swarmkit/api"
-	"github.com/docker/swarmkit/cmd/swarmctl/common"
+	"github.com/docker/swarmkit/examples/swarmctl/common"
 	"github.com/spf13/cobra"
 )
