+++ examples/swarmctl/task/cmd.go
