+++ examples/swarmctl/task/print.go
@@ -7,7 +7,7 @@ import (
 	"text/tabwriter"
 
 	"github.com/docker/swarmkit/api"
-	"github.com/docker/swarmkit/cmd/swarmctl/common"
+	"github.com/docker/swarmkit/examples/swarmctl/common"
 	"github.com/docker/swarmkit/protobuf/ptypes"
 )
