+++ examples/swarmd/main.go
