+++ tools/swarm-bench/collector.go
